wip: common library compiles now
